Code cleanup

+++ scripts/test.sh
@@ -3,21 +3,36 @@ echo "Start evaluation task queues"
 
 # Hyperparameters
 name_array=("eth" "hotel" "univ" "zara1" "zara2")
+device_id_array=(0 1 2 3 4)
 prefix="graph-tern_"
 suffix="_experiment"
 
-device_id_array=(0 1 2 3 4)
-PID_array=()
+# Arguments
+while getopts p:s:d:i: flag
+do
+  case "${flag}" in
+    p) prefix=${OPTARG};;
+    s) suffix=${OPTARG};;
+    d) dataset_array=(${OPTARG});;
+    i) device_id_array=(${OPTARG});;
+    *) echo "usage: $0 [-p PREFIX] [-s SUFFIX] [-d \"eth hotel univ zara1 zara2\"] [-i \"0 1 2 3 4\"]" >&2
+      exit 1 ;;
+  esac
+done
 
-if [ $# -eq 5 ]; then
- echo "CUDA_VISIBLE_DEVICES set to: $1 $2 $3 $4 $5"
- device_id_array=($1 $2 $3 $4 $5)
+if [ ${#dataset_array[@]} -ne ${#device_id_array[@]} ]
+then
+    printf "Arrays must all be same length. "
+    printf "len(dataset_array)=${#dataset_array[@]} and len(device_id_array)=${#device_id_array[@]}\n"
+    exit 1
 fi
 
 # Signal handler
+PID_array=()
+
 sighdl ()
 {
-  echo "Kill training processes"
+  echo "Kill evaluation processes"
   for (( i=0; i<${#name_array[@]}; i++ ))
   do
     kill ${PID_array[$i]}
@@ -36,10 +51,6 @@ do
   --tag "${prefix}""${name_array[$i]}""${suffix}" &
   PID_array[$i]=$!
   printf " job ${#PID_array[@]} pid ${PID_array[$i]}\n"
-done
-
-for (( i=0; i<${#name_array[@]}; i++ ))
-do
   wait ${PID_array[$i]}
 done
